full ethereum functionality with ledger using module separation
