+++ packages/bitcore-hardware/examples/ledger/address.ts
@@ -1,8 +1,11 @@
 import 'source-map-support/register.js';
-import Ledger from '../../src/ledger.js';
+import Ledger from '../../src/ledger/wallet.js';
+
+const args = process.argv.slice(2);
+const chain = args[0] || 'BTC';
 
 const ledger = new Ledger();
 await ledger.connect();
-console.log(await ledger.getAddress({ index: 0 }));
+console.log(await ledger.getAddress({ chain }));
 await ledger.disconnect();
 process.exit(0);
