+++ packages/bitcore-hardware/src/burner.ts
@@ -4,6 +4,7 @@ import { NFC } from 'nfc-pcsc';
 import { Base } from './base.js';
 import { DataType } from './types/burnerTypes.js';
 import { BaseParams } from './types/paramTypes.js';
+import { EveryUtxoType, TransactionType, UtxoType } from './types/txTypes.js';
 
 const { Address, PublicKey, crypto } = BitcoreLib;
 
@@ -244,43 +245,6 @@ export default class Burner implements Base {
     return data.firmwareVersion[index].value;
   }
 }
-
-/** Transaction data that can be converted into a Transaction via Transaction(tx) */
-type TransactionType = BitcoreLib.Transaction | string | Buffer | object;
-
-/**
- * Standard utxo type use for internal processing.
- * Property names are from bitcore-lib's UnspentOutput.
- * Note, UnspentOutput addresses and scripts are Address and Script classes respectively,
- * here they are both strings.
- */
-export type UtxoType = {
-  txId: string;
-  outputIndex: number;
-  satoshis: number;
-  script: string;
-  address?: string;
-};
-
-/** 
- * Utxo type for functions were the received utxo type is unknown.
- * Could either be in the format of UnspentOutput, UnspentOutput.toObject, or from bitcore-node.
- */
-export type EveryUtxoType = Partial<UtxoType & {
-  // bitcore-node specific properties
-  mintTxid: string;
-  mintIndex: number;
-  mintHeight: number;
-  value: number;
-  // UnspentOutput.toObject specific properties
-  txid: string;
-  amount: number;
-  vout: number;
-  scriptPubKey: string;
-  // UnspentOutput specific, allow for Script and Address objects
-  script: string | BitcoreLib.Script;
-  address: string | BitcoreLib.Address;
-}>;
 
 type CommandType = Record<string, any> & {
   name: string;
