+++ packages/bitcore-hardware/src/ledger/dmk.ts
