+++ packages/bitcore-hardware/src/ledger/modules/index.ts
@@ -1,0 +1,7 @@
+import BitcoinModule from './btc.js';
+import EthereumModule from './eth.js';
+
+export {
+  BitcoinModule,
+  EthereumModule
+};
